message-analyzer/requirements.txt
confluent_kafka==2.6.0
psycopg2_binary==2.9.10
transformers==4.46.2
torch==2.3.1+cpu